requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
google-api-python-client==2.179.0
google-auth==2.40.3
websockets==15.0.1
